Make settings.ts the entry point for all `settings.*.ts` files.
