+++ packages/client/src/settings/index.test.ts
@@ -1,0 +1,7 @@
+import * as index from './index';
+
+describe('settings/index', () => {
+    test('', () => {
+        expect(typeof index.enableLocaleForTarget).toBe('function');
+    });
+});
